Fixed type error

+++ app/src/main/java/com/example/campusguessr/SettingsActivity.java
@@ -86,7 +86,7 @@ public class SettingsActivity extends AppCompatActivity {
           else {
             currentUser.setDesiredDifficulty(50);
           }
-          currentUser.setScore(user.get("score"));
+          currentUser.setScore((Long)user.get("score"));
           Log.d(TAG, "onComplete: " + currentUser);
           NewUsername.setText(currentUser.getName());
           DesiredDistance.setProgress(currentUser.getDesiredDistance());
